pyproject.toml
[tool.poetry]
name = "healthecho-chatbot"
version = "0.1.0"
description = "A conversational AI chatbot for health-related queries."
authors = ["Vijay Kumar Kodam, vijay.kodam98@gmail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.115.8"
uvicorn = "0.34.0"
groq = "0.16.0"
pydantic = "2.10.6"
python-dotenv = "1.0.1"
requests = "2.32.3"
langchain-community = "0.3.16"
langchain-groq = "0.2.4"
langchain-core = "0.3.33"
transformers = "4.48.2"
langchain-huggingface = "0.1.2"
sentence-transformers = "3.4.1"
torch = "2.6.0"
torchvision = "0.21.0"
torchaudio = "2.6.0"
google-cloud-firestore = "2.20.0"
beautifulsoup4 = "4.13.1"
flask = "3.1.0"
langdetect = "1.0.9"
google-cloud-aiplatform = "1.79.0"
google-cloud-storage = "2.19.0"
faiss-cpu = "1.10.0"
flask-cors = "5.0.0"
markdown = "3.7"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
